Extract functions into separate file
